# application/src/main/java/com/cuentas/backend/domain/Category.java
package com.cuentas.backend.domain;

import java.time.LocalDateTime;

public class Category {
    private Long categoryId;
    private Long userId;
    private String name;
    private String description;
    private String type; // income | expense
    private LocalDateTime createdAt;
    private LocalDateTime updatedAt;

    // Getters y setters
    public Long getCategoryId() { return categoryId; }
    public void setCategoryId(Long categoryId) { this.categoryId = categoryId; }
    public Long getUserId() { return userId; }
    public void setUserId(Long userId) { this.userId = userId; }
    public String getName() { return name; }
    public void setName(String name) { this.name = name; }
    public String getDescription() { return description; }
    public void setDescription(String description) { this.description = description; }
    public String getType() { return type; }
    public void setType(String type) { this.type = type; }
    public LocalDateTime getCreatedAt() { return createdAt; }
    public void setCreatedAt(LocalDateTime createdAt) { this.createdAt = createdAt; }
    public LocalDateTime getUpdatedAt() { return updatedAt; }
    public void setUpdatedAt(LocalDateTime updatedAt) { this.updatedAt = updatedAt; }
}